Add middleware for student user profile checks
Implemented a new StudentCheck middleware to ensure that student details (student_id and github) are set before allowing lesson and workout access. If these details are not set, the user is redirected to profile page. Additionally, made 'student_id' and 'github' fields nullable in the database migrations and added an informative alert message on the profile page.