Made 'url' field nullable and updated StudentWorkout model and controller
Made the 'url' attribute on the 'student_workouts' table nullable to allow for the optional submission of a URL. Additionally, we now enable mass assignment for select fields in the StudentWorkout model. It also creates a StudentWorkout entry when a LessonWorkout is mounted, and the submitted url is saved to this entry in the database.