Added save confirmation message and fixed URL duplication
Updated the student lesson workout view and method in the Livewire component. Added a confirmation message to indicate to the user when the workout details have been saved. Also removed a duplicate URL that was causing redundancy and instead dispatched a 'saved' event after the student workout URL is saved.