Add checker_user_id to student_workouts table
This commit adds a new nullable integer column "checker_user_id" to the "student_workouts" table that stores the ID of the user who checks the students' workouts. The authenticated user's ID is now stored when workouts are checked or unchecked. It also changes the list of admin IDs in config/personal.php.