Refactor mail sending logic to use full workout object
The mail sending logic in the student's lesson workout has been amended to pass the full workout object, rather than just the title. This makes the approach more flexible, as it will be able to include more data in the mail if needed in the future. Consequently, the NewWorkoutSubmitMail class was adjusted to accept this object and extract the title when necessary.